refactor: split types/json_convert.go — temporal and geospatial converters into separate files
